Use form struct to input and output data
