+++ api/model/article_model.go
@@ -2,9 +2,9 @@ package model
 
 import (
 	"context"
-	"encoding/json"
 	"github.com/kentaro-m/spider/api/entity"
 	"github.com/kentaro-m/spider/api/repository"
+	"github.com/kentaro-m/spider/api/form"
 	"github.com/satori/go.uuid"
 	"golang.org/x/xerrors"
 	"net/http"
@@ -12,8 +12,8 @@ import (
 )
 
 type ArticleModel interface {
-	Get(ctx context.Context) ([]*entity.Article, error)
-	Create(ctx context.Context, r *http.Request) error
+	Get(ctx context.Context, g *form.GetArticleForm) ([]*entity.Article, error)
+	Create(ctx context.Context, r *http.Request, c *form.CreateArticleForm) error
 }
 
 func NewArticleModel(r repository.ArticleRepository) ArticleModel {
@@ -26,8 +26,24 @@ type articleModel struct {
 	repo repository.ArticleRepository
 }
 
-func (a articleModel) Get(ctx context.Context) ([]*entity.Article, error) {
-	payload, err := a.repo.Get(ctx)
+func (a articleModel) Get(ctx context.Context, g *form.GetArticleForm) ([]*entity.Article, error) {
+	if g.Limit == 0 {
+		g.Limit = 50
+	}
+
+	if g.Sort == "" {
+		g.Sort = "desc"
+	}
+
+	if g.Until.IsZero() {
+		g.Until = time.Now().UTC().In(time.FixedZone("Asia/Tokyo", 9*60*60))
+	}
+
+	if g.Since.IsZero() {
+		g.Since = time.Now().Add(- time.Hour * 24 * 7).UTC().In(time.FixedZone("Asia/Tokyo", 9*60*60))
+	}
+
+	payload, err := a.repo.Get(ctx, g)
 
 	if err != nil {
 		return nil, xerrors.Errorf("failed to get articles from DB: %w", err)
@@ -36,22 +52,19 @@ func (a articleModel) Get(ctx context.Context) ([]*entity.Article, error) {
 	return payload, err
 }
 
-func (a articleModel) Create(ctx context.Context, r *http.Request) error {
+func (a articleModel) Create(ctx context.Context, r *http.Request, c *form.CreateArticleForm) error {
 	timeStamp := time.Now().UTC().In(time.FixedZone("Asia/Tokyo", 9*60*60))
 
 	article := entity.Article{
 		ID:        uuid.NewV4().String(),
+		Title: c.Title,
+		URL: c.URL,
+		PubDate: c.PubDate,
 		CreatedAt: timeStamp,
 		UpdatedAt: timeStamp,
 	}
 
-	err := json.NewDecoder(r.Body).Decode(&article)
-
-	if err != nil {
-		return xerrors.Errorf("failed to read JSON-encoded value: %w", err)
-	}
-
-	err = a.repo.Create(r.Context(), &article)
+	err := a.repo.Create(r.Context(), &article)
 
 	if err != nil {
 		return xerrors.Errorf("failed to insert an article to DB: %w", err)
